+++ src/server/routes/pessoas.routes.ts
@@ -2,6 +2,8 @@ import { Router } from 'express';
 
 import { ensureAuthenticated } from '../shared/middleware';
 import { PessoasController } from '../controllers';
+import { ensureResourceOwner } from '../shared/middleware/EnsureResourceOwner';
+import { ETableNames } from '../database/ETableNames';
 
 const pessoasRouter = Router();
 
@@ -29,12 +31,14 @@ pessoasRouter
     .put(
         '/pessoas/:id',
         ensureAuthenticated,
+        ensureResourceOwner(ETableNames.pessoa),
         PessoasController.updateByIdValidation,
         PessoasController.updateById
     )
     .delete(
         '/pessoas/:id',
         ensureAuthenticated,
+        ensureResourceOwner(ETableNames.pessoa),
         PessoasController.deleteByIdValidation,
         PessoasController.deleteById
     );
